% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusivity_table.R
\name{exclusivity_table}
\alias{exclusivity_table}
\title{Calculate Exclusivity Table}
\usage{
exclusivity_table(Vars, DF)
}
\arguments{
\item{Vars}{Character vector of variable names for which exclusivity should be calculated.}

\item{DF}{Data frame containing the dummy variables and 'Area_sq_Km' column.}
}
\value{
A data frame summarizing exclusivity statistics for the specified variables.
}
\description{
Given a data frame with dummy variables (encoded as "Yes" and NA), and an 'Area_sq_Km' column, this
function calculates exclusivity statistics for specified variables.
}
\examples{

# Create a sample data frame 'LT'
LT <- data.frame(
  planned_strictly_protected = c("Yes", "Yes", "Yes", "Yes", NA, NA, NA, NA),
  strictly_protected = c("Yes", "Yes", NA, NA, "Yes", "Yes", NA, NA),
  sea_natural_areas = c("Yes", NA, "Yes", NA, "Yes", NA, "Yes", NA),
  Area_sq_Km = c(4163.9369, 0.0043, 136.5007, 0.0001, 2153.2234, 23.0762, 24540.5325, 327940.2363)
)

# Calculate exclusivity table
exclusivity_table(DF = LT, Vars = c("strictly_protected", "sea_natural_areas"))

}
