% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subclassgenerator.R
\name{OpenGenerator}
\alias{OpenGenerator}
\title{Generate "Open" Class}
\usage{
OpenGenerator(DF)
}
\arguments{
\item{DF}{The input DataFrame to be modified.}
}
\value{
A modified DataFrame with the "Open" class assigned in the `Subclass` column.
}
\description{
This function checks a DataFrame and classifies rows as "Open" based on specific criteria.
}
\examples{

# Load required libraries if not already loaded
# library(your_library_name)

# Create a sample DataFrame (DF)
DF <- data.frame(
  p3_klit = c("Eng", "Hede", "Mose", "Overdrev", NA, NA),
  Urort_Skov = c(NA, NA, NA, NA, "Untouched", "Untouched"),
  Forest = c(NA, NA, NA, NA, NA, NA),
  markblokkort = c(NA, NA, NA, NA, NA, NA),
  National_Parks = c(NA, NA, NA, NA, NA, NA),
  Subclass = c(NA, NA, NA, NA, NA, NA)
)

# Apply the OpenGenerator function to classify rows as "Open"
DF <- OpenGenerator(DF)

}
