% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderAndTHresholdData.R
\name{orderAndThresholdData}
\alias{orderAndThresholdData}
\title{Order and Threshold Data}
\usage{
orderAndThresholdData(data, threshold, priority_order)
}
\arguments{
\item{data}{A data frame containing the input dataset.}

\item{threshold}{A numeric value representing the threshold for cumulative area.}

\item{priority_order}{A character vector specifying the priority order for processing data.}
}
\value{
A data frame containing the filtered data based on the threshold.
}
\description{
This function orders a dataset based on a priority order, calculates the cumulative area, and filters the data
to meet a specified threshold.
}
\details{
The function iterates through the specified priority order, processes each priority in sequence, and
stops once the cumulative area exceeds the threshold. It ensures that IDs processed in earlier priorities are
not considered in subsequent priorities.
}
\examples{
# Load your dataset (replace 'your_data' with the actual variable name of your dataset)
your_data <- read.csv("your_dataset.csv")

# Set the threshold for cumulative area (e.g., 5 square kilometers)
threshold <- 5

# Define your priority order (customize it as needed)
custom_priority_order <- c("DIQ", "DOQ", "RVB", "RVE", "LAV", "BUF")

# Call the function with the specified priority order
filtered_dataset <- orderAndThresholdData(your_data, threshold, custom_priority_order)



}
